chore: divided cside module into submodules
